+++ src/app/app-routing.module.ts
@@ -6,7 +6,7 @@ import { SponsorsComponent } from './sponsors/sponsors.component';
 import { NewsComponent } from './news/news.component';
 import { ContactUsComponent } from './contact-us/contact-us.component';
 import { TeamComponent } from './team/team.component';
-import { MainComponent } from './main/main.component';
+import { MainComponent } from './business/main/main.component';
 import { ArticleComponent } from './news/article/article.component';
 import { ElysiaComponent } from './elysia/elysia.component';
